experimental/draftRss/requirements.txt
asyncpg
feedparser
httpx
pyyaml